Add new icons component file
Added a new file, icons.tsx, to the components directory. This file imports a variety of icons for use throughout the app, exports an Icon type and a collection of all available icons. The aim is to centralize all icons to ensure consistency and ease of use across the app.